fixing releaseDate and adding softwareVersion
releaseDate:
this key contains the date at which the latest version was released. This date is mandatory if the software has been released at least once and thus the version number is present.

softwareVersion:
This key contains the latest stable version number of the software. The version number is a string that is not meant to be interpreted and parsed but just displayed; parsers should not assume semantic versioning or any other specific version format.
The key can be omitted if the software is currently in initial development and has never been released yet.

More info: https://docs.italia.it/italia/developers-italia/publiccodeyml-en/en/master/schema.core.html